# swagger.yaml
openapi: 3.0.0
info:
  title: Warehouse Co
  description: Warehouse and Inventory management system restful Api
  contact: {}
  version: '1.0'
servers:
- url: https://odd-rose-turkey-wear.cyclic.app/api/v1
  variables: {}
paths:
...
  /registerUser:
    post:
      tags:
      - Auth
      summary: Register User
      description: Register other users besides like the admin, cashier and warehouse-manager
      operationId: RegisterUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/RegisterUserRequest'
              - example:
                  name: Umoru Emmanuel`
                  role: ADMIN
                  password: '123456789'
                  email: umoru@gmail.com
            example:
              name: Umoru Emmanuel`
              role: ADMIN
              password: '123456789'
              email: umoru@gmail.com
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    RegisterUserRequest:
      title: RegisterUserRequest
      required:
      - name
      - role
      - password
      - email
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        password:
          type: string
        email:
          type: string
      example:
        name: Umoru Emmanuel`
        role: ADMIN
        password: '123456789'
        email: umoru@gmail.com
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...